+++ src/0_storage.js
@@ -14,8 +14,11 @@ var BRANCH_KEY_PREFIX = 'BRANCH_WEBSDK_KEY';
 /** @typedef {{get:function(string), set:function(string, (string|boolean)), remove:function(string), clear:function(), isEnabled:function()}} */
 var storage;
 
-/** @typedef {{listProperties: function(), setString: function({string}, {string}), getString: function({string})}}*/
-Ti.App.Properties;
+// Try catch so tests pass
+try {
+	/** @typedef {{listProperties: function(), setString: function({string}, {string}), getString: function({string})}}*/
+	Ti.App.Properties;
+} catch(e) {}
 
 /**
  * @class BranchStorage