+++ test/0_storage.js
@@ -1,5 +1,7 @@
 goog.require('storage'); // jshint ignore:line
 
+var BRANCH_KEY_PREFIX = 'BRANCH_WEBSDK_KEY';
+
 describe('session storage', function() {
 	var storage = new BranchStorage(['session']); // jshint ignore:line
 
@@ -9,7 +11,7 @@ describe('session storage', function() {
 
 	it('should set an item', function() {
 		storage.set(itemKey, itemValue);
-		assert.equal(sessionStorage.getItem(itemKey), itemValue, 'key / vaue stored');
+		assert.equal(sessionStorage.getItem(BRANCH_KEY_PREFIX + itemKey), itemValue, 'key / vaue stored');
 	});
 
 	it('shold get stored item with key', function() {
@@ -45,7 +47,7 @@ describe('local storage', function() {
 
 	it('should set an item in localStorage', function() {
 		storage.set(itemKey, itemValue);
-		assert.equal(localStorage.getItem(itemKey), itemValue, 'key / vaue stored');
+		assert.equal(localStorage.getItem(BRANCH_KEY_PREFIX + itemKey), itemValue, 'key / vaue stored');
 	});
 
 	it('shold get stored item with key', function() {